Fix overflow when reporting stats for reads longer than 255 bp (#12)
* fix overflow when reporting stats for reads longer than 255 bp (#11)

* revert typo

* fix type change missed in 40270e2

+++ src/combine_counts.rs
@@ -8,7 +8,7 @@ use std::string::String;
 pub(crate) struct Count {
     read12: u8,
     orientation: u8,
-    read_pos: u8,
+    read_pos: u32,
     bq_bin: u8,
     context: [char; 2],
     homopolymer_dist: i8,
@@ -46,7 +46,7 @@ impl Count {
             read_pos: sp
                 .next()
                 .expect("not enough columns in line {s}")
-                .parse::<u8>()
+                .parse::<u32>()
                 .expect("error parsing int"),
             context: {
                 let mut ctx = sp
